Decimate signal

Downsamples by an integer factor with anti-aliasing filter. An 80%-Nyquist lowpass Butterworth filter is applied before decimation to prevent aliasing.

Usage

decimate(x, factor, filter_order = 8L, output_assay = "decimated")

Arguments

x

A PhysioExperiment object.

factor

Integer decimation factor.

filter_order

Order of the anti-aliasing lowpass filter.

output_assay

Name for the output assay.

Value

A new PhysioExperiment object with sampling rate equal to the original rate divided by factor. The time dimension is reduced accordingly. Column data and metadata are carried over.

References

Crochiere, R.E. & Rabiner, L.R. (1983). "Multirate Digital Signal Processing." Prentice Hall.

See also